Refactored to multiple components
