+++ src/components/CalculatorDisplay.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CalculatorDisplay = () => (
+  <>
+    <h2>Your Math Magician Calculator</h2>
+    <input type="text" placeholder={0} disabled />
+  </>
+
+);
+
+export default CalculatorDisplay;
