+++ src/styles/Calculator.css
@@ -38,7 +38,7 @@ input[type="text"]::placeholder {
 button {
   padding: 1rem;
   outline: none;
-  border: 0.5px solid hsl(210deg 12.5% 96.86%);
+  border: 0.5px solid hsl(236deg 6.55% 55.1%);
   font-family: 'Poppins', sans-serif;
   font-size: 1rem;
 }
